refactor(core): consolidate interactions/ and DOM utils into controllers/
Moves all DOM-side-effect code into the new controllers/ directory.
Pure helpers stay in utils/. This realises the layer boundary defined
in Section 1 of the v2.0.0 design spec.

- Moves packages/core/src/interactions/* into packages/core/src/controllers/
- Moves packages/core/src/utils/portal.ts into controllers/
- Moves packages/core/src/utils/suggestionContainer.ts into controllers/
- Updates packages/core/src/index.ts to re-export from controllers/index.ts
- Keeps utils/index.ts re-exporting the moved symbols for source-level
  compatibility within v2.0.0. Section 6 cleans up the dual export.
